+++ docs/fr_FR/changelog.md
@@ -4,9 +4,12 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte
 
+
+
 # 19/01/2022
 - encrypte les infos de configuration dans la BDD
 - masque les paramètres d'appel du démon
+- correction message de log démon
 
 # 18/01/2022
 - contrôle des dépendances installées